Refactor layout and modal components, update styles
