+++ package.json
@@ -19,7 +19,7 @@
     "@uniswap/v3-periphery": "^1.4.4",
     "@uniswap/v3-sdk": "^3.26.0",
     "antd": "^5.27.6",
-    "axios": "^1.12.2",
+    "axios": "^1.13.0",
     "dayjs": "^1.11.18",
     "ethers": "^6.15.0",
     "graphql-request": "^7.3.1",
@@ -28,15 +28,15 @@
     "i18next": "^25.6.0",
     "i18next-http-backend": "^3.0.2",
     "lodash": "^4.17.21",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0",
-    "react-i18next": "^16.2.0",
+    "react": "^18.3.0",
+    "react-dom": "^18.3.0",
+    "react-i18next": "^16.2.1",
     "react-responsive": "^10.0.1",
     "react-router-dom": "^7.9.4",
     "styled-components": "^6.1.19",
     "uuid": "^13.0.0",
-    "viem": "^2.38.4",
-    "wagmi": "^2.18.2"
+    "viem": "^2.38.5",
+    "wagmi": "^2.19.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.38.0",
@@ -45,8 +45,8 @@
     "@fontsource/roboto-mono": "^5.2.8",
     "@types/lodash": "^4.17.20",
     "@types/node": "^24.9.1",
-    "@types/react": "^19.2.2",
-    "@types/react-dom": "^19.2.2",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^5.1.0",
     "autoprefixer": "^10.4.21",
     "csstype": "^3.1.3",
