--- src/components/swap-settings/index.scss
@@ -1,285 +1,0 @@
-@use "~variables" as *;
-
-.swap-settings {
-  background-color: $background-secondary;
-  border-radius: rem(20);
-  display: flex;
-  flex-direction: column;
-  gap: rem(16);
-  padding: rem(16);
-
-  .heading {
-    align-items: center;
-    display: flex;
-    flex-direction: row-reverse;
-    height: rem(34);
-    justify-content: space-between;
-    position: relative;
-
-    .text {
-      font-size: rem(22);
-      font-weight: 500;
-      left: 50%;
-      margin-top: rem(-2);
-      position: absolute;
-      top: 50%;
-      transform: translate(-50%, -50%);
-    }
-
-    .toggle {
-      cursor: pointer;
-      font-size: 24px;
-    }
-
-    .reset {
-      color: $button-primary;
-      cursor: pointer;
-      font-weight: 600;
-      text-transform: uppercase;
-    }
-  }
-
-  .slippage {
-    display: flex;
-    flex-direction: column;
-    gap: rem(12);
-
-    .ant-input-number-input {
-      height: rem(36);
-    }
-
-    .ant-input-number-affix-wrapper {
-      border-radius: rem(12);
-      width: 100%;
-    }
-
-    .label {
-      font-weight: 600;
-    }
-
-    .item {
-      background-color: $background-tertiary;
-      border-radius: rem(12);
-      align-items: center;
-      cursor: pointer;
-      display: flex;
-      font-size: rem(12);
-      font-weight: 500;
-      height: rem(36);
-      justify-content: center;
-      width: calc(100% / 3 - rem(8));
-    }
-
-    .active {
-      background-color: $button-secondary;
-    }
-
-    .list {
-      display: flex;
-      gap: rem(12);
-    }
-  }
-
-  .modes {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-
-    .label {
-      font-weight: 600;
-      line-height: rem(20);
-    }
-
-    .ant-radio {
-      display: none;
-    }
-
-    .ant-radio-wrapper {
-      align-items: center;
-      border-radius: rem(20);
-      cursor: pointer;
-      display: flex;
-      font-size: rem(12);
-      font-weight: 500;
-      height: rem(32);
-      justify-content: center;
-      width: rem(88);
-    }
-
-    .ant-radio-wrapper-checked {
-      background-color: $button-secondary;
-    }
-
-    .ant-radio-group {
-      background-color: $background-tertiary;
-      border-radius: rem(20);
-      display: flex;
-      gap: rem(4);
-      padding: rem(4);
-    }
-  }
-
-  .advanced-mode {
-    display: none;
-    flex-direction: column;
-    gap: rem(16);
-
-    .ant-input-number-input {
-      height: rem(56);
-    }
-
-    .ant-input-number {
-      border-radius: rem(12);
-      width: 100%;
-    }
-
-    .ant-input-number-affix-wrapper {
-      border-radius: rem(12);
-      width: 100%;
-    }
-
-    .ant-form-item-additional {
-      display: none;
-    }
-
-    .ant-form-item {
-      margin: 0;
-    }
-
-    &.active {
-      display: flex;
-    }
-  }
-
-  .basic-mode {
-    display: none;
-
-    .icon {
-      font-size: 24px;
-      position: relative;
-      stroke: $blue-persian;
-    }
-
-    .title {
-      display: flex;
-      flex: 1;
-      flex-direction: column;
-      position: relative;
-
-      .text {
-        font-size: rem(16);
-        font-weight: 600;
-        text-transform: uppercase;
-      }
-
-      .speed {
-        font-weight: 500;
-      }
-    }
-
-    .gwei {
-      position: relative;
-    }
-
-    .ant-radio {
-      display: none;
-    }
-
-    .ant-radio-label {
-      align-items: center;
-      display: flex;
-      flex: 1;
-      gap: rem(12);
-      padding: rem(12);
-    }
-
-    .ant-radio-wrapper {
-      border: solid rem(1) $border-normal;
-      border-radius: rem(12);
-      overflow: hidden;
-      position: relative;
-      transition: all 0.2s;
-
-      &::before {
-        bottom: 0;
-        content: "";
-        left: 0;
-        opacity: 0.2;
-        position: absolute;
-        right: 0;
-        top: 0;
-        z-index: 0;
-      }
-
-      &:hover {
-        border-color: $button-secondary;
-      }
-    }
-
-    .ant-radio-wrapper-checked {
-      border-color: $button-secondary;
-
-      &::before {
-        background-color: $blue-persian;
-      }
-    }
-
-    .ant-radio-group {
-      display: flex;
-      flex-direction: column;
-      gap: rem(16);
-    }
-
-    &.active {
-      display: block;
-    }
-  }
-
-  .button {
-    height: rem(38);
-  }
-}
-
-@media (min-width: rem(768)) {
-  .swap-settings {
-    .heading {
-      height: rem(44);
-
-      .text {
-        font-size: rem(28);
-      }
-    }
-
-    .slippage {
-      display: flex;
-      flex-direction: row;
-      flex-wrap: wrap;
-
-      .ant-input-number-affix-wrapper {
-        width: calc(50% - rem(6));
-      }
-
-      .label {
-        width: 100%;
-      }
-
-      .list {
-        width: calc(50% - rem(6));
-
-        li {
-          cursor: pointer;
-        }
-      }
-    }
-
-    .basic-mode {
-      .gwei {
-        font-size: rem(16);
-      }
-
-      .item {
-        padding: rem(16);
-      }
-    }
-  }
-}
