+++ src/components/swap-vult/index.scss
@@ -1,59 +1,6 @@
 @use "~variables" as *;
 
-@keyframes rotate {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
 .swap-vult {
-  background-color: $background-secondary;
-  border-radius: rem(20);
-  display: flex;
-  flex-direction: column;
-  gap: rem(16);
-  padding: rem(16);
-  .refresh {
-    margin-left: auto;
-    cursor: pointer;
-
-    &.spinning {
-      animation: rotate 1s linear infinite;
-    }
-  }
-  .heading {
-    align-items: center;
-    display: flex;
-    flex-direction: row-reverse;
-    height: rem(34);
-    justify-content: space-between;
-    position: relative;
-
-    .text {
-      font-size: rem(22);
-      font-weight: 500;
-      left: 50%;
-      margin-top: rem(-2);
-      position: absolute;
-      top: 50%;
-      transform: translate(-50%, -50%);
-    }
-
-    .toggle {
-      cursor: pointer;
-    }
-
-    .reset {
-      color: $button-primary;
-      cursor: pointer;
-      font-weight: 600;
-      text-transform: uppercase;
-    }
-  }
-
   .info {
     display: flex;
     flex-direction: column;
@@ -74,35 +21,6 @@
     .success {
       color: $button-primary;
     }
-  }
-
-  .whitelist {
-    align-items: center;
-    display: flex;
-    border-radius: rem(12);
-    font-weight: 500;
-    gap: rem(8);
-    height: rem(50);
-    padding: 0 rem(16);
-
-    svg {
-      height: rem(16);
-      width: rem(16);
-    }
-  }
-
-  .islisted {
-    background-color: #042436;
-    border: solid rem(1) #13c89d40;
-    color: $button-primary;
-
-    svg {
-      stroke: $button-primary;
-    }
-  }
-
-  .notlisted {
-    border: solid rem(1) $border-light;
   }
 
   .swap {
