+++ src/icons/CrossIcon.tsx
@@ -1,0 +1,17 @@
+import { FC, SVGProps } from "react";
+
+export const CrossIcon: FC<SVGProps<SVGSVGElement>> = (props) => (
+  <svg
+    fill="none"
+    height="1em"
+    stroke="currentColor"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    strokeWidth="1.5"
+    viewBox="0 0 24 24"
+    width="1em"
+    {...props}
+  >
+    <path d="M18 6L6 18M6 6L18 18" />
+  </svg>
+);
