--- src/layouts/default/index.scss
@@ -1,55 +1,0 @@
-@use "~variables" as *;
-@use "~animations" as *;
-
-.default-layout {
-  min-height: 100vh;
-
-  .ant-layout-header {
-    .ant-tabs-content-holder {
-      display: none;
-    }
-
-    .ant-tabs-nav {
-      &::before {
-        content: none;
-      }
-    }
-
-    .ant-tabs-tab-btn {
-      a {
-        display: block;
-        font-weight: 500;
-        line-height: rem(64);
-        padding: 0 rem(20);
-      }
-    }
-
-    .ant-tabs-tab {
-      padding: 0;
-    }
-
-    .ant-tabs-tab-active {
-      a {
-        color: $blue-persian;
-      }
-    }
-
-    .ant-tabs {
-      flex-grow: 1;
-    }
-  }
-
-  .ant-layout-content {
-    margin: 0 auto;
-    padding: rem(16);
-    width: 100%;
-  }
-}
-
-@media (min-width: rem(768)) {
-  .default-layout {
-    .ant-layout-content {
-      padding: rem(24) rem(16);
-    }
-  }
-}
