--- src/styles/_animations.scss
@@ -1,22 +1,0 @@
-// Global animations for reuse across components
-@use "~variables" as *;
-
-// Wiggle animation keyframes
-@keyframes wiggle {
-  0% { transform: translateX(0); }
-  50% { transform: translateX(3px); }
-  75% { transform: translateX(-2px); }
-  90% { transform: translateX(1px); }
-  100% { transform: translateX(0); }
-}
-
-// Tab indicator mixin
-@mixin tab-indicator {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  height: rem(2);
-  width: 50%;
-  background-color: $alert-success;
-  transition: transform 0.3s ease;
-}
