+++ src/styles/_normalize.scss
@@ -48,7 +48,3 @@ svg {
 video {
   display: block;
 }
-
-body {
-  user-select: none;
-}
